Rename generated snippets so filename == region tag
